@@ -1,6 +1,6 @@
 {
   "name": "@apidevtools/json-schema-ref-parser",
-  "version": "8.0.0",
+  "version": "9.0.6",
   "description": "Parse, Resolve, and Dereference JSON Schema $ref pointers",
   "keywords": [
     "json",
@@ -42,7 +42,7 @@
     "test": "npm run test:node && npm run test:typescript && npm run test:browser && npm run lint",
     "test:node": "mocha",
     "test:browser": "karma start --single-run",
-    "test:typescript": "tsc --noEmit --strict --lib esnext test/specs/typescript-definition.spec.ts",
+    "test:typescript": "tsc --noEmit --strict --lib esnext,dom test/specs/typescript-definition.spec.ts",
     "coverage": "npm run coverage:node && npm run coverage:browser",
     "coverage:node": "nyc node_modules/mocha/bin/mocha",
     "coverage:browser": "npm run test:browser -- --coverage",
@@ -51,26 +51,27 @@
     "release": "npm run upgrade && npm run clean && npm test && npm run bump"
   },
   "devDependencies": {
-    "@babel/polyfill": "^7.7.0",
-    "@jsdevtools/eslint-config-modular": "^8.0.0",
-    "@jsdevtools/host-environment": "^2.0.2",
-    "@jsdevtools/karma-config": "^3.1.2",
-    "@jsdevtools/version-bump-prompt": "^6.0.2",
+    "@babel/polyfill": "^7.10.4",
+    "@jsdevtools/eslint-config": "^1.0.7",
+    "@jsdevtools/host-environment": "^2.0.4",
+    "@jsdevtools/karma-config": "^3.1.7",
+    "@jsdevtools/version-bump-prompt": "^6.0.6",
     "@types/json-schema": "^7.0.4",
-    "@types/node": "^13.1.2",
+    "@types/node": "^14.0.25",
     "chai": "^4.2.0",
-    "eslint": "^6.8.0",
-    "karma": "^4.4.1",
+    "chai-subset": "^1.6.0",
+    "eslint": "^7.5.0",
+    "karma": "^5.0.2",
     "karma-cli": "^2.0.0",
-    "mocha": "^7.1.0",
+    "mocha": "^8.0.1",
     "npm-check": "^5.9.0",
-    "nyc": "^15.0.0",
+    "nyc": "^15.0.1",
     "shx": "^0.3.2",
-    "typescript": "^3.7.4"
+    "typescript": "^3.9.7"
   },
   "dependencies": {
+    "@jsdevtools/ono": "^7.1.3",
     "call-me-maybe": "^1.0.1",
-    "js-yaml": "^3.13.1",
-    "@jsdevtools/ono": "^7.1.0"
+    "js-yaml": "^3.13.1"
   }
 }
