@@ -11,10 +11,10 @@ module.exports = karmaConfig({
   sourceDir: "lib",
   fixtures: "test/fixtures/**/*.js",
   browsers: {
-    chrome: !host.os.windows,
-    firefox: host.os.linux,
-    safari: host.os.linux,    // SauceLabs
-    edge: host.os.linux,      // SauceLabs
-    ie: host.os.windows,
+    chrome: host.ci ? host.os.linux : true,
+    firefox: host.ci ? host.os.linux : true,
+    safari: host.ci ? host.os.linux : host.os.mac,    // SauceLabs in CI
+    edge: host.ci ? host.os.linux : host.os.windows,  // SauceLabs in CI
+    ie: host.ci ? host.os.windows : false,  // IE needs to run by itself, due to Babel transforms
   },
 });
