@@ -1,5 +1,7 @@
 "use strict";
 
+const { ParserError } = require("../util/errors");
+
 let TEXT_REGEXP = /\.(txt|htm|html|md|xml|js|min|map|css|scss|less|svg)$/i;
 
 module.exports = {
@@ -48,7 +50,7 @@ module.exports = {
    * @param {string} file.url       - The full URL of the referenced file
    * @param {string} file.extension - The lowercased file extension (e.g. ".txt", ".html", etc.)
    * @param {*}      file.data      - The file contents. This will be whatever data type was returned by the resolver
-   * @returns {Promise<string>}
+   * @returns {string}
    */
   parse (file) {
     if (typeof file.data === "string") {
@@ -58,7 +60,7 @@ module.exports = {
       return file.data.toString(this.encoding);
     }
     else {
-      throw new Error("data is not text");
+      throw new ParserError("data is not text", file.url);
     }
   }
 };
