@@ -1,6 +1,7 @@
 "use strict";
 
-const YAML = require("../util/yaml");
+const { ParserError } = require("../util/errors");
+const yaml = require("js-yaml");
 
 module.exports = {
   /**
@@ -36,20 +37,23 @@ module.exports = {
    * @param {*}      file.data      - The file contents. This will be whatever data type was returned by the resolver
    * @returns {Promise}
    */
-  parse (file) {
-    return new Promise(((resolve, reject) => {
-      let data = file.data;
-      if (Buffer.isBuffer(data)) {
-        data = data.toString();
-      }
+  async parse (file) {      // eslint-disable-line require-await
+    let data = file.data;
+    if (Buffer.isBuffer(data)) {
+      data = data.toString();
+    }
 
-      if (typeof data === "string") {
-        resolve(YAML.parse(data));
+    if (typeof data === "string") {
+      try {
+        return yaml.safeLoad(data);
       }
-      else {
-        // data is already a JavaScript value (object, array, number, null, NaN, etc.)
-        resolve(data);
+      catch (e) {
+        throw new ParserError(e.message, file.url);
       }
-    }));
+    }
+    else {
+      // data is already a JavaScript value (object, array, number, null, NaN, etc.)
+      return data;
+    }
   }
 };
