@@ -4,7 +4,7 @@ module.exports = Pointer;
 
 const $Ref = require("./ref");
 const url = require("./util/url");
-const { ono } = require("@jsdevtools/ono");
+const { JSONParserError, InvalidPointerError, MissingPointerError, isHandledError } = require("./util/errors");
 const slashes = /\//g;
 const tildes = /~/g;
 const escapedSlash = /~1/g;
@@ -64,6 +64,7 @@ function Pointer ($ref, path, friendlyPath) {
  *
  * @param {*} obj - The object that will be crawled
  * @param {$RefParserOptions} options
+ * @param {string} pathFromRoot - the path of place that initiated resolving
  *
  * @returns {Pointer}
  * Returns a JSON pointer whose {@link Pointer#value} is the resolved value.
@@ -71,20 +72,26 @@ function Pointer ($ref, path, friendlyPath) {
  * the {@link Pointer#$ref} and {@link Pointer#path} will reflect the resolution path
  * of the resolved value.
  */
-Pointer.prototype.resolve = function (obj, options) {
-  let tokens = Pointer.parse(this.path);
+Pointer.prototype.resolve = function (obj, options, pathFromRoot) {
+  let tokens = Pointer.parse(this.path, this.originalPath);
 
   // Crawl the object, one token at a time
-  this.value = obj;
+  this.value = unwrapOrThrow(obj);
+
   for (let i = 0; i < tokens.length; i++) {
     if (resolveIf$Ref(this, options)) {
       // The $ref path has changed, so append the remaining tokens to the path
       this.path = Pointer.join(this.path, tokens.slice(i));
     }
 
+    if (typeof this.value === "object" && this.value !== null && "$ref" in this.value) {
+      return this;
+    }
+
     let token = tokens[i];
-    if (this.value[token] === undefined) {
-      throw ono.syntax(`Error resolving $ref pointer "${this.originalPath}". \nToken "${token}" does not exist.`);
+    if (this.value[token] === undefined || this.value[token] === null) {
+      this.value = null;
+      throw new MissingPointerError(token, this.originalPath);
     }
     else {
       this.value = this.value[token];
@@ -92,7 +99,10 @@ Pointer.prototype.resolve = function (obj, options) {
   }
 
   // Resolve the final value
-  resolveIf$Ref(this, options);
+  if (!this.value || this.value.$ref && url.resolve(this.path, this.value.$ref) !== pathFromRoot) {
+    resolveIf$Ref(this, options);
+  }
+
   return this;
 };
 
@@ -117,7 +127,8 @@ Pointer.prototype.set = function (obj, value, options) {
   }
 
   // Crawl the object, one token at a time
-  this.value = obj;
+  this.value = unwrapOrThrow(obj);
+
   for (let i = 0; i < tokens.length - 1; i++) {
     resolveIf$Ref(this, options);
 
@@ -150,9 +161,10 @@ Pointer.prototype.set = function (obj, value, options) {
  * {@link https://tools.ietf.org/html/rfc6901#section-3}
  *
  * @param {string} path
+ * @param {string} [originalPath]
  * @returns {string[]}
  */
-Pointer.parse = function (path) {
+Pointer.parse = function (path, originalPath) {
   // Get the JSON pointer from the path's hash
   let pointer = url.getHash(path).substr(1);
 
@@ -171,7 +183,7 @@ Pointer.parse = function (path) {
   }
 
   if (pointer[0] !== "") {
-    throw ono.syntax(`Invalid $ref pointer "${pointer}". Pointers must begin with "#/"`);
+    throw new InvalidPointerError(pointer, originalPath === undefined ? path : originalPath);
   }
 
   return pointer.slice(1);
@@ -222,7 +234,7 @@ function resolveIf$Ref (pointer, options) {
       pointer.circular = true;
     }
     else {
-      let resolved = pointer.$ref.$refs._resolve($refPath, options);
+      let resolved = pointer.$ref.$refs._resolve($refPath, pointer.path, options);
       pointer.indirections += resolved.indirections + 1;
 
       if ($Ref.isExtended$Ref(pointer.value)) {
@@ -264,7 +276,16 @@ function setValue (pointer, token, value) {
     }
   }
   else {
-    throw ono.syntax(`Error assigning $ref pointer "${pointer.path}". \nCannot set "${token}" of a non-object.`);
+    throw new JSONParserError(`Error assigning $ref pointer "${pointer.path}". \nCannot set "${token}" of a non-object.`);
   }
+  return value;
+}
+
+
+function unwrapOrThrow (value) {
+  if (isHandledError(value)) {
+    throw value;
+  }
+
   return value;
 }
