@@ -41,7 +41,7 @@ function $Refs () {
  * @param {...string|string[]} [types] - Only return paths of the given types ("file", "http", etc.)
  * @returns {string[]}
  */
-$Refs.prototype.paths = function (types) {
+$Refs.prototype.paths = function (types) {    // eslint-disable-line no-unused-vars
   let paths = getPaths(this._$refs, arguments);
   return paths.map((path) => {
     return path.decoded;
@@ -54,7 +54,7 @@ $Refs.prototype.paths = function (types) {
  * @param {...string|string[]} [types] - Only return references of the given types ("file", "http", etc.)
  * @returns {object}
  */
-$Refs.prototype.values = function (types) {
+$Refs.prototype.values = function (types) {   // eslint-disable-line no-unused-vars
   let $refs = this._$refs;
   let paths = getPaths($refs, arguments);
   return paths.reduce((obj, path) => {
@@ -79,7 +79,7 @@ $Refs.prototype.toJSON = $Refs.prototype.values;
  */
 $Refs.prototype.exists = function (path, options) {
   try {
-    this._resolve(path, options);
+    this._resolve(path, "", options);
     return true;
   }
   catch (e) {
@@ -95,7 +95,7 @@ $Refs.prototype.exists = function (path, options) {
  * @returns {*} - Returns the resolved value
  */
 $Refs.prototype.get = function (path, options) {
-  return this._resolve(path, options).value;
+  return this._resolve(path, "", options).value;
 };
 
 /**
@@ -139,11 +139,12 @@ $Refs.prototype._add = function (path) {
  * Resolves the given JSON reference.
  *
  * @param {string} path - The path being resolved, optionally with a JSON pointer in the hash
+ * @param {string} pathFromRoot - The path of `obj` from the schema root
  * @param {$RefParserOptions} [options]
  * @returns {Pointer}
  * @protected
  */
-$Refs.prototype._resolve = function (path, options) {
+$Refs.prototype._resolve = function (path, pathFromRoot, options) {
   let absPath = url.resolve(this._root$Ref.path, path);
   let withoutHash = url.stripHash(absPath);
   let $ref = this._$refs[withoutHash];
@@ -152,7 +153,7 @@ $Refs.prototype._resolve = function (path, options) {
     throw ono(`Error resolving $ref pointer "${path}". \n"${withoutHash}" not found.`);
   }
 
-  return $ref.resolve(absPath, options, path);
+  return $ref.resolve(absPath, options, path, pathFromRoot);
 };
 
 /**
