@@ -4,6 +4,7 @@ const $Ref = require("./ref");
 const Pointer = require("./pointer");
 const parse = require("./parse");
 const url = require("./util/url");
+const { isHandledError } = require("./util/errors");
 
 module.exports = resolveExternal;
 
@@ -53,7 +54,7 @@ function resolveExternal (parser, options) {
 function crawl (obj, path, $refs, options) {
   let promises = [];
 
-  if (obj && typeof obj === "object") {
+  if (obj && typeof obj === "object" && !ArrayBuffer.isView(obj)) {
     if ($Ref.isExternal$Ref(obj)) {
       promises.push(resolve$Ref(obj, path, $refs, options));
     }
@@ -101,11 +102,25 @@ async function resolve$Ref ($ref, path, $refs, options) {
   }
 
   // Parse the $referenced file/url
-  const result = await parse(resolvedPath, $refs, options);
+  try {
+    const result = await parse(resolvedPath, $refs, options);
 
-  // Crawl the parsed value
-  // console.log('Resolving $ref pointers in %s', withoutHash);
-  let promises = crawl(result, withoutHash + "#", $refs, options);
+    // Crawl the parsed value
+    // console.log('Resolving $ref pointers in %s', withoutHash);
+    let promises = crawl(result, withoutHash + "#", $refs, options);
 
-  return Promise.all(promises);
+    return Promise.all(promises);
+  }
+  catch (err) {
+    if (!options.continueOnError || !isHandledError(err)) {
+      throw err;
+    }
+
+    if ($refs._$refs[withoutHash]) {
+      err.source = url.stripHash(path);
+      err.path = url.safePointerToPath(url.getHash(path));
+    }
+
+    return [];
+  }
 }
