@@ -2,6 +2,7 @@
 const fs = require("fs");
 const { ono } = require("@jsdevtools/ono");
 const url = require("../util/url");
+const { ResolverError } = require("../util/errors");
 
 module.exports = {
   /**
@@ -40,7 +41,7 @@ module.exports = {
         path = url.toFileSystemPath(file.url);
       }
       catch (err) {
-        reject(ono.uri(err, `Malformed URI: ${file.url}`));
+        reject(new ResolverError(ono.uri(err, `Malformed URI: ${file.url}`), file.url));
       }
 
       // console.log('Opening file: %s', path);
@@ -48,7 +49,7 @@ module.exports = {
       try {
         fs.readFile(path, (err, data) => {
           if (err) {
-            reject(ono(err, `Error opening file "${path}"`));
+            reject(new ResolverError(ono(err, `Error opening file "${path}"`), path));
           }
           else {
             resolve(data);
@@ -56,7 +57,7 @@ module.exports = {
         });
       }
       catch (err) {
-        reject(ono(err, `Error opening file "${path}"`));
+        reject(new ResolverError(ono(err, `Error opening file "${path}"`), path));
       }
     }));
   }
