@@ -4,6 +4,7 @@ const http = require("http");
 const https = require("https");
 const { ono } = require("@jsdevtools/ono");
 const url = require("../util/url");
+const { ResolverError } = require("../util/errors");
 
 module.exports = {
   /**
@@ -106,8 +107,8 @@ function download (u, httpOptions, redirects) {
         }
         else if (res.statusCode >= 300) {
           if (redirects.length > httpOptions.redirects) {
-            reject(ono({ status: res.statusCode },
-              `Error downloading ${redirects[0]}. \nToo many redirects: \n  ${redirects.join(" \n  ")}`));
+            reject(new ResolverError(ono({ status: res.statusCode },
+              `Error downloading ${redirects[0]}. \nToo many redirects: \n  ${redirects.join(" \n  ")}`)));
           }
           else if (!res.headers.location) {
             throw ono({ status: res.statusCode }, `HTTP ${res.statusCode} redirect with no location header`);
@@ -123,7 +124,7 @@ function download (u, httpOptions, redirects) {
         }
       })
       .catch((err) => {
-        reject(ono(err, `Error downloading ${u.href}`));
+        reject(new ResolverError(ono(err, `Error downloading ${u.href}`), u.href));
       });
   }));
 }
