@@ -49,7 +49,9 @@ describe("HTTP options", () => {
     }
   });
 
-  describe("http.redirect", () => {
+  // 2020-07-08 - The HTTPBin redirect endpoints are suddenly returning 404 errors. Not sure why 🤷‍♂️
+  // TODO: Re-enable these tests once HTTPBin is working again
+  describe.skip("http.redirect", () => {
     if (host.browser.safari && host.karma && host.karma.ci) {
       // These tests fail in Safari when running on Sauce Labs (they pass when running on Safari locally).
       // It gets an XHR error when trying to reach httpbin.org.
@@ -155,7 +157,7 @@ describe("HTTP options", () => {
       let schema = await parser.parse("https://petstore.swagger.io/v2/swagger.json");
 
       expect(schema).to.be.an("object");
-      expect(schema).not.to.be.empty;
+      expect(schema).not.to.be.empty;  // eslint-disable-line no-unused-expressions
       expect(parser.schema).to.equal(schema);
     });
 
@@ -169,7 +171,7 @@ describe("HTTP options", () => {
       });
 
       expect(schema).to.be.an("object");
-      expect(schema).not.to.be.empty;
+      expect(schema).not.to.be.empty;  // eslint-disable-line no-unused-expressions
       expect(parser.schema).to.equal(schema);
     });
 
@@ -186,7 +188,7 @@ describe("HTTP options", () => {
 
           // The request succeeded, which means this browser doesn't support CORS.
           expect(schema).to.be.an("object");
-          expect(schema).not.to.be.empty;
+          expect(schema).not.to.be.empty;  // eslint-disable-line no-unused-expressions
           expect(parser.schema).to.equal(schema);
         }
         catch (err) {
