@@ -1,11 +1,15 @@
 "use strict";
 
-const { expect } = require("chai");
+const chai = require("chai");
+const chaiSubset = require("chai-subset");
+chai.use(chaiSubset);
+const { expect } = chai;
 const $RefParser = require("../../..");
 const helper = require("../../utils/helper");
 const path = require("../../utils/path");
 const parsedSchema = require("./parsed");
 const dereferencedSchema = require("./dereferenced");
+const { ResolverError, UnmatchedResolverError, JSONParserErrorGroup } = require("../../../lib/util/errors");
 
 describe("options.resolve", () => {
   it('should not resolve external links if "resolve.external" is disabled', async () => {
@@ -46,7 +50,7 @@ describe("options.resolve", () => {
           // A custom resolver for "foo://" URLs
           foo: {
             canRead: /^foo\:\/\//i,
-            read (file) {
+            read (_file) {
               return { bar: { baz: "hello world" }};
             }
           }
@@ -62,7 +66,7 @@ describe("options.resolve", () => {
           // A custom resolver for "foo://" URLs
           foo: {
             canRead: /^foo\:\/\//i,
-            read (file, callback) {
+            read (_file, callback) {
               callback(null, { bar: { baz: "hello world" }});
             }
           }
@@ -79,7 +83,7 @@ describe("options.resolve", () => {
             // A custom resolver for "foo://" URLs
             foo: {
               canRead: /^foo\:\/\//i,
-              read (file) {
+              read (_file) {
                 return Promise.resolve({ bar: { baz: "hello world" }});
               }
             }
@@ -97,7 +101,7 @@ describe("options.resolve", () => {
           badResolver: {
             order: 1,
             canRead: true,
-            read (file) {
+            read (_file) {
               throw new Error("BOMB!!!");
             }
           },
@@ -111,4 +115,73 @@ describe("options.resolve", () => {
     expect(schema).to.deep.equal(dereferencedSchema);
   });
 
+  it("should normalize errors thrown by resolvers", async () => {
+    try {
+      await $RefParser.dereference({ $ref: path.abs("specs/resolvers/resolvers.yaml") }, {
+        resolve: {
+          // A custom resolver that always fails
+          file: {
+            order: 1,
+            canRead: true,
+            parse () {
+              throw new Error("Woops");
+            }
+          }
+        }
+      });
+      helper.shouldNotGetCalled();
+    }
+    catch (err) {
+      expect(err).to.be.instanceof(ResolverError);
+      expect(err.message).to.contain("Error opening file");
+    }
+  });
+
+  it("should throw an error if no resolver returned a result", async () => {
+    try {
+      await $RefParser.dereference(path.rel("specs/resolvers/resolvers.yaml"), {
+        resolve: {
+          http: false,
+          file: {
+            order: 1,
+            canRead: true,
+            read () {
+
+            }
+          }
+        }
+      });
+      helper.shouldNotGetCalled();
+    }
+    catch (err) {
+      // would time out otherwise
+      expect(err).to.be.an.instanceOf(ResolverError);
+    }
+  });
+
+  it("should throw a grouped error if no resolver can be matched and continueOnError is true", async () => {
+    const parser = new $RefParser();
+    try {
+      await parser.dereference(path.abs("specs/resolvers/resolvers.yaml"), {
+        resolve: {
+          file: false,
+          http: false,
+        },
+        continueOnError: true,
+      });
+      helper.shouldNotGetCalled();
+    }
+    catch (err) {
+      expect(err).to.be.instanceof(JSONParserErrorGroup);
+      expect(err.errors.length).to.equal(1);
+      expect(err.errors).to.containSubset([
+        {
+          name: UnmatchedResolverError.name,
+          message: message => message.startsWith("Could not find resolver for"),
+          path: [],
+          source: message => message.endsWith("specs/resolvers/resolvers.yaml"),
+        },
+      ]);
+    }
+  });
 });
